Merge commits from origin

+++ src/Microsoft.PowerShell.Commands.Utility/commands/utility/StartSleepCommand.cs
@@ -18,7 +18,7 @@ namespace Microsoft.PowerShell.Commands
 
         #region IDisposable
         /// <summary>
-        ///  Dispose method of IDisposable interface.
+        /// Dispose method of IDisposable interface.
         /// </summary>
         public void Dispose()
         {
@@ -29,6 +29,7 @@ namespace Microsoft.PowerShell.Commands
                     _waitHandle.Dispose();
                     _waitHandle = null;
                 }
+
                 _disposed = true;
             }
         }
@@ -38,15 +39,15 @@ namespace Microsoft.PowerShell.Commands
         #region parameters
 
         /// <summary>
-        /// Allows sleep time to be specified in seconds
+        /// Allows sleep time to be specified in seconds.
         /// </summary>
         [Parameter(Position = 0, Mandatory = true, ParameterSetName = "Seconds", ValueFromPipeline = true,
                    ValueFromPipelineByPropertyName = true)]
-        [ValidateRangeAttribute(0, int.MaxValue / 1000)]
-        public int Seconds { get; set; }
+        [ValidateRangeAttribute(0.0, (double)(int.MaxValue / 1000))]
+        public double Seconds { get; set; }
 
         /// <summary>
-        /// Allows sleep time to be specified in milliseconds
+        /// Allows sleep time to be specified in milliseconds.
         /// </summary>
         [Parameter(Mandatory = true, ParameterSetName = "Milliseconds", ValueFromPipelineByPropertyName = true)]
         [ValidateRangeAttribute(0, int.MaxValue)]
@@ -57,19 +58,18 @@ namespace Microsoft.PowerShell.Commands
 
         #region methods
 
-        //Wait handle which is used by thread to sleep.
+        // Wait handle which is used by thread to sleep.
         private ManualResetEvent _waitHandle;
 
-        //object used for synchronizes pipeline thread and stop thread
-        //access to waitHandle
+        // object used for synchronizes pipeline thread and stop thread
+        // access to waitHandle
         private object _syncObject = new object();
 
-        //this is set to true by stopProcessing
+        // this is set to true by stopProcessing
         private bool _stopping = false;
 
         /// <summary>
-        /// This method causes calling thread to sleep for
-        /// specified milliseconds
+        /// This method causes calling thread to sleep for specified milliseconds.
         /// </summary>
         private void Sleep(int milliSecondsToSleep)
         {
@@ -80,6 +80,7 @@ namespace Microsoft.PowerShell.Commands
                     _waitHandle = new ManualResetEvent(false);
                 }
             }
+
             if (_waitHandle != null)
             {
                 _waitHandle.WaitOne(milliSecondsToSleep, true);
@@ -87,7 +88,7 @@ namespace Microsoft.PowerShell.Commands
         }
 
         /// <summary>
-        ///
+        /// ProcessRecord method.
         /// </summary>
         protected override void ProcessRecord()
         {
@@ -96,7 +97,7 @@ namespace Microsoft.PowerShell.Commands
             switch (ParameterSetName)
             {
                 case "Seconds":
-                    sleepTime = Seconds * 1000;
+                    sleepTime = (int)(Seconds * 1000);
                     break;
 
                 case "Milliseconds":
@@ -109,14 +110,12 @@ namespace Microsoft.PowerShell.Commands
             }
 
             Sleep(sleepTime);
-        } // EndProcessing
+        }
 
         /// <summary>
-        /// stopprocessing override
+        /// StopProcessing override.
         /// </summary>
-        protected override
-        void
-        StopProcessing()
+        protected override void StopProcessing()
         {
             lock (_syncObject)
             {
@@ -129,6 +128,5 @@ namespace Microsoft.PowerShell.Commands
         }
 
         #endregion
-    } // StartSleepCommand
-} // namespace Microsoft.PowerShell.Commands
-
+    }
+}
